// ui/scheduler_ui/src/api/models/shiftResponse.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Scheduler API
 * OpenAPI spec version: 1.0.0
 */

export interface ShiftResponse {
  id: string;
  schedule_id: string;
  shift_template_id: string;
  name: string;
  start_time: string;
  end_time: string;
  created_at: string;
  updated_at: string;
}